assignment/Digital_Employee_PoC/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-google-genai
requests
pydantic
